Update Logic.java
Added at the end of the Logic facade class the create project function. There is a small (could be extended) assert calls for the data found in a projectAttributes. I stuck with the projectName as a simple variable to test, but we can add more or choose a different variable from the attributes class.