Create ProjectsDb.java
Included additional ProjectsDb class that will extend the abstract EntitiesDb class. Right now I noted that the "createEntity" is not utilized in the ProjectsDb but rather in the EntitiesDb. The StudentsDb does Override the createEntities, but it is due to a needed step of putting into a document; meaning it still calls the super class EntitesDb to process the createEntity.